Rearranged project for better encapsulation. Added new classes to improve the fluent API experience. Added a feature to make specifying a fail message mandatory. Improved the README.md.
