--- src/main/java/org/solidcoding/validation/ChainingPredicate.java
@@ -1,7 +1,0 @@
-package org.solidcoding.validation;
-
-public interface ChainingPredicate<T, R> {
-
-  R and(T second);
-
-}
