--- src/main/java/org/solidcoding/validation/DefineThat.java
@@ -1,13 +1,0 @@
-package org.solidcoding.validation;
-
-import java.util.function.Predicate;
-
-public class DefineThat {
-
-  private DefineThat() {}
-
-  public static <T> Predicate<T> itShould(Predicate<T> rule) {
-    return rule;
-  }
-
-}
