+++ src/main/java/org/solidcoding/validation/api/ChainingPredicate.java
@@ -1,0 +1,7 @@
+package org.solidcoding.validation.api;
+
+public interface ChainingPredicate<T, R> {
+
+    R and(T second);
+
+}
