+++ src/main/java/org/solidcoding/validation/api/Rule.java
@@ -1,7 +1,7 @@
-package org.solidcoding.validation;
+package org.solidcoding.validation.api;
 
 import java.util.function.Predicate;
 
 public interface Rule<T> extends Predicate<T> {
-  String getFailMessage();
+    String getFailMessage();
 }
