--- src/test/java/org/solidcoding/validation/testutil/TestValue.java
@@ -1,6 +1,0 @@
-package org.solidcoding.validation.testutil;
-
-public class TestValue {
-
-  public static final String TEST_CONTENT = "test";
-}
