+++ src/test/java/testutil/TestObject.java
@@ -1,4 +1,4 @@
-package org.solidcoding.validation.testutil;
+package testutil;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,20 +9,20 @@ public class TestObject {
 
     private final List<String> testList = new ArrayList<>();
 
-    public void add(String value) {
-        testList.add(value);
-    }
-
-    public boolean hasValues() {
-        return !testList.isEmpty();
-    }
-
     public TestObject() {
         this.message = "Nothing to report";
     }
 
     public TestObject(String message) {
         this.message = message;
+    }
+
+    public void add(String value) {
+        testList.add(value);
+    }
+
+    public boolean hasValues() {
+        return !testList.isEmpty();
     }
 
     public String getMessage() {
