+++ src/test/java/testutil/TestValue.java
@@ -1,0 +1,6 @@
+package testutil;
+
+public class TestValue {
+
+    public static final String TEST_CONTENT = "test";
+}
